Update Actix3 example
