+++ examples/actix3_example/api/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "actix3-example-api"
+version = "0.1.0"
+authors = ["Sam Samai <sam@studio2pi.com.au>"]
+edition = "2021"
+publish = false
+
+[dependencies]
+actix3-example-core = { path = "../core" }
+actix-http = "2"
+actix-web = "3"
+actix-flash = "0.2"
+actix-files = "0.5"
+futures = { version = "^0.3" }
+futures-util = { version = "^0.3" }
+tera = "1.8.0"
+dotenv = "0.15"
+listenfd = "0.3.3"
+serde = "1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+entity = { path = "../entity" }
+migration = { path = "../migration" }
