+++ examples/actix3_example/api/static/images/favicon.png
