+++ examples/actix3_example/api/templates/index.html.tera
